Allow Python 2
Previously, Python 3 was forced to be required because we'd like to use
shutil.which() to check if the command "pdflatex" is available in the path.
We now have to allow (and encourage) Python 2 instead because
the package PyPDF2 apparently has incompatibility issues with Python 3,
especially for PDF files with .eps vector images. Until this issue with
PyPDF2 is solved, we probably should use Python 2.